+++ core/js/src/main/scala/diplomacy/variant/standard/rule/JsStandardDiplomacyUnit.scala
@@ -1,7 +1,10 @@
 package diplomacy.variant.standard.rule
 
+import scala.scalajs.js.annotation.{ JSExport, JSExportAll }
+
 import diplomacy.board.Power
 
+@JSExport @JSExportAll
 case class JsStandardDiplomacyUnit[Power_ <: Power](unit: DiplomacyUnit[Power_]) {
   def hold() = Order.Hold(unit)
   def move(destination: Location[Power_]) = Order.Move(unit, destination, false)
