+++ core/shared/src/main/scala/diplomacy/board/MilitaryBranch.scala
@@ -1,5 +1,8 @@
 package diplomacy.board
 
+import scala.scalajs.js.annotation.{ JSExportAll, JSExportDescendentObjects }
+
+@JSExportDescendentObjects @JSExportAll
 trait MilitaryBranch {
   val name: Name
   override def toString: String = this.name.toString
