+++ core/shared/src/main/scala/diplomacy/board/Name.scala
@@ -1,5 +1,8 @@
 package diplomacy.board
 
+import scala.scalajs.js.annotation.{ JSExport, JSExportAll }
+
+@JSExport @JSExportAll
 final case class Name(name: String, abbreviatedName: String) {
   def this(name: String) = this(name, name)
