+++ core/shared/src/main/scala/diplomacy/board/Power.scala
@@ -1,5 +1,8 @@
 package diplomacy.board
 
+import scala.scalajs.js.annotation.{ JSExportDescendentObjects, JSExportAll }
+
+@JSExportDescendentObjects @JSExportAll
 trait Power {
   val name: String
   override def toString: String = this.name
