+++ core/shared/src/main/scala/diplomacy/rule/InvalidOrderMessage.scala
@@ -1,3 +1,6 @@
 package diplomacy.rule
 
+import scala.scalajs.js.annotation.{ JSExport, JSExportAll }
+
+@JSExport @JSExportAll
 case class InvalidOrderMessage(message: String)
