+++ core/shared/src/main/scala/diplomacy/variant/standard/board/Implicits.scala
@@ -1,0 +1,9 @@
+package diplomacy.variant.standard.board
+
+object Implicits {
+  /* Turn */
+  implicit class Year2Turn(year: Int) {
+    val Spring = Turn(year, Season.Spring)
+    val Autumn = Turn(year, Season.Autumn)
+  }
+}
