+++ core/shared/src/main/scala/diplomacy/variant/standard/board/Keywords.scala
@@ -1,0 +1,10 @@
+package diplomacy.variant.standard.board
+
+import scala.scalajs.js.annotation.{ JSExport, JSExportAll }
+
+@JSExport @JSExportAll
+object Keywords {
+  /* Season */
+  val Spring = Season.Spring
+  val Autumn = Season.Autumn
+}
