+++ core/shared/src/main/scala/diplomacy/variant/standard/board/Season.scala
@@ -1,5 +1,8 @@
 package diplomacy.variant.standard.board
 
+import scala.scalajs.js.annotation.JSExportDescendentObjects
+
+@JSExportDescendentObjects
 sealed abstract class Season(name: String) {
   override def toString: String = name
 }
