+++ core/shared/src/main/scala/diplomacy/variant/standard/board/Turn.scala
@@ -1,8 +1,11 @@
 package diplomacy.variant.standard.board
 
+import scala.scalajs.js.annotation.{ JSExportAll, JSExport }
+
 import diplomacy.variant.standard.rule.{ Turn => BaseTurn }
 import diplomacy.variant.standard.board.Keywords._
 
+@JSExportAll
 case class Turn(year: Int, season: Season) extends BaseTurn {
   override def toString: String = s"${year}-${season}"
   val isBuildable: Boolean = season == Autumn
@@ -10,6 +13,7 @@ case class Turn(year: Int, season: Season) extends BaseTurn {
 }
 
 object Turn {
+  @JSExport
   def nextTurn(turn: Turn): Turn = turn match {
     case Turn(year, Spring) => Turn(year, Autumn)
     case Turn(year, Autumn) => Turn(year + 1, Spring)
