+++ core/shared/src/main/scala/diplomacy/variant/standard/map/packge.scala
@@ -1,0 +1,11 @@
+package diplomacy.variant.standard
+
+import diplomacy.board.{ Name, Province }
+import diplomacy.variant.standard.rule._
+import diplomacy.variant.standard.rule.Keywords._
+import diplomacy.variant.standard.map.Keywords._
+import diplomacy.util.LabeledUndirectedGraph
+
+package object map {
+  val map = DiplomacyMap.map
+}
