+++ core/shared/src/main/scala/diplomacy/variant/standard/rule/Order.scala
@@ -1,17 +1,23 @@
 package diplomacy.variant.standard.rule
 
+import scala.scalajs.js.annotation.{ JSExport, JSExportAll, JSExportDescendentObjects }
+
 import diplomacy.board.{ Power, Location => BaseLocation }
 import diplomacy.rule.{ Order => BaseOrder }
 
+@JSExportDescendentObjects
 sealed trait Order[Power_ <: Power] extends BaseOrder[Power_, MilitaryBranch]
 sealed trait MovementOrder[Power_ <: Power] extends Order[Power_]
 
+@JSExport
 object Order {
+  @JSExportAll
   final case class Hold[Power_ <: Power](unit: DiplomacyUnit[Power_])
     extends MovementOrder[Power_] {
     override def toString: String = s"${this.unit} H"
   }
 
+  @JSExportAll
   final case class Move[Power_ <: Power](
     unit: DiplomacyUnit[Power_], destination: Location[Power_], useConvoy: Boolean
   ) extends MovementOrder[Power_] {
@@ -21,6 +27,7 @@ object Order {
     }
   }
 
+  @JSExportAll
   final case class Support[Power_ <: Power](
     unit: DiplomacyUnit[Power_], target: Either[Hold[Power_], Move[Power_]]
   ) extends MovementOrder[Power_] with BaseLocation.TypeHelper {
@@ -42,6 +49,7 @@ object Order {
     }
   }
 
+  @JSExportAll
   final case class Convoy[Power_ <: Power](
     unit: DiplomacyUnit[Power_], target: Move[Power_]
   ) extends MovementOrder[Power_] {
@@ -50,17 +58,20 @@ object Order {
 
   sealed trait NotMovementOrder[Power_ <: Power] extends Order[Power_]
 
+  @JSExportAll
   final case class Retreat[Power_ <: Power](
     unit: DiplomacyUnit[Power_], destination: Location[Power_]
   ) extends NotMovementOrder[Power_] {
     override def toString: String = s"${unit} R ${this.destination}"
   }
 
+  @JSExportAll
   final case class Disband[Power_ <: Power](unit: DiplomacyUnit[Power_])
     extends NotMovementOrder[Power_] {
     override def toString: String = s"Disband: ${unit}"
   }
 
+  @JSExportAll
   final case class Build[Power_ <: Power](unit: DiplomacyUnit[Power_])
     extends NotMovementOrder[Power_] {
     //require(Option(unit.power) == unit.location.province.homeOf)
