+++ core/shared/src/main/scala/diplomacy/variant/standard/rule/Phase.scala
@@ -1,6 +1,10 @@
 package diplomacy.variant.standard.rule
 
+import scala.scalajs.js.annotation.{ JSExport, JSExportDescendentObjects }
+
+@JSExport
 object Phase {
+  @JSExportDescendentObjects
   sealed abstract class Phase(name: String) {
     override def toString: String = name
   }
