+++ core/shared/src/main/scala/diplomacy/variant/standard/rule/ProvinceStatus.scala
@@ -1,5 +1,8 @@
 package diplomacy.variant.standard.rule
 
+import scala.scalajs.js.annotation.{ JSExportDescendentObjects, JSExportAll }
+
+@JSExportDescendentObjects @JSExportAll
 sealed abstract class ProvinceStatus(name: String) {
   override def toString: String = name
 }
