+++ core/shared/src/main/scala/diplomacy/variant/standard/rule/Result.scala
@@ -1,7 +1,10 @@
 package diplomacy.variant.standard.rule
 
+import scala.scalajs.js.annotation.{ JSExportDescendentObjects, JSExportAll }
+
 import diplomacy.board.{ Power, Province }
 
+@JSExportDescendentObjects @JSExportAll
 sealed abstract class Result(name: String) {
   override def toString: String = name
 }
@@ -9,6 +12,7 @@ sealed abstract class Result(name: String) {
 object Result {
   object Success extends Result("Success")
   object Failed extends Result("Failed")
+  @JSExportAll
   case class Dislodged[Power_ <: Power](attackedFrom: Province[Power_])
     extends Result(s"Dislodged from ${attackedFrom.name}")
   object Bounced extends Result("Bounced")
