+++ core/shared/src/main/scala/diplomacy/variant/standard/rule/Turn.scala
@@ -1,5 +1,8 @@
 package diplomacy.variant.standard.rule
 
+import scala.scalajs.js.annotation.{ JSExportDescendentObjects, JSExportAll }
+
+@JSExportDescendentObjects @JSExportAll
 trait Turn {
   val isBuildable: Boolean
   val isOccupationUpdateable: Boolean
