+++ core/shared/src/main/scala/diplomacy/variant/standard/rule/UnitStatus.scala
@@ -1,7 +1,10 @@
 package diplomacy.variant.standard.rule
 
+import scala.scalajs.js.annotation.{ JSExportDescendentObjects, JSExportAll }
+
 import diplomacy.board.{ Power, Province }
 
+@JSExportDescendentObjects @JSExportAll
 sealed abstract class UnitStatus[Power_ <: Power](name: String) {
   override def toString: String = name
 }
